requirements.txt
flask == 2.0.3
mysql-connector-python == 8.0.28
dotenv
gunicorn
